requirements.txt
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.18
boto==2.34.0
celery==3.1.17
certifi==14.05.14
Collectfast==0.2.1
colorama==0.3.2
django-admin-export==2.0
django-appconf==0.6
django-apptemplates==0.0.1
django-autocomplete-light==3.2.9
django-bower==5.0.1
django-cached-field==1.2.15
django-celery==3.2.1
django-ckeditor==5.3.0
django-compressor==2.1.1
django-constance==2.0.0
git+https://github.com/jnm/django-file-import@f9078d9751060b0c6adfdd772fc89c7e32c0cd99
django-custom-field==2.9
django-dajax-ng==0.9.4
django-dajaxice==0.7
django-daterange-filter==1.0.0
django-extensions==1.4.6
django-filter==0.8
django-floppy-gumby==0.1
django-floppyforms==1.7.0
django-grappelli==2.10.1
django-heroku-redisify==0.2.1
django-impersonate==0.9
django-ldap-groups-bsc==0.1.5
django-libsass==0.2
django-localflavor==1.0
django-mass-edit==3.0
django-pagination==1.0.7
django-picklefield==1.0.0
django-positions==0.5.1
django-redis-cache==0.13.0
django-report-builder==3.1.2
django-report-scaffold==0.1.4
django-report-utils==0.3.1
django-responsive-dashboard==1.1
django-reversion==2.0.9
django-rosetta==0.7.4
django-simple-import==2.0.2
#django-storages-redux==1.2.1
django-storages
django-tenant-schemas==1.5.2
django-toolbelt==0.0.1
django-widget-tweaks==1.3
djangorestframework-bulk==0.1.3
djangorestframework==3.0.5
faulthandler==2.4
flower==0.7.3
html5lib==0.999
httpagentparser==1.7.5
jdcal==1.3
jsonfield==1.0.0
Markdown==2.5.1
medusa==0.5.4
meld3==1.0.0
newrelic==2.38.0.31
oauthlib==0.7.2
odfpy==0.9.6
openpyxl==2.1.2
polib==1.0.5
poster==0.8.1
PyJWT==0.3.0
python-dateutil==2.2
python-ldap==2.4.18
python-openid==2.2.5
python-social-auth==0.3.6
social-auth-app-django==1.2.0
pytz==2014.9
raven==5.1.1
reportlab==3.1.44
requests-oauthlib==0.4.2
requests==2.4.3
simplejson==3.6.5
six==1.8.0
suds==0.4
supervisor==3.1.3
tenant-schemas-celery==0.1.1
urllib3==1.9.1
uWSGI==2.0.15
wsgiref==0.1.2
xlrd==0.9.3
xlwt==0.7.5
git+https://github.com/jnm/django-file-import@f9078d9751060b0c6adfdd772fc89c7e32c0cd99
#django-hstore
